+++ sources/translationfile.cpp
@@ -4,6 +4,6 @@ TranslationFile::TranslationFile() = default;
 
 void TranslationFile::setName(std::string name) { SceneName = name; }
 std::string TranslationFile::getName() { return SceneName; }
-int TranslationFile::getNbFunctions() const { return FunctionsInFile.size(); }
+int TranslationFile::getNbFunctions() const { return static_cast<int>(FunctionsInFile.size()); }
 
 void TranslationFile::addFunction(const function& fun) { FunctionsInFile.push_back(fun); }
